hashtab: Check for vacant slots even in the event of no hash match
Should the subsequent 32 slots be occupied by values with non-matching
hashes, the implementation would just assume that the first such
slot was vacant. This is not necessarily the case. Instead of assuming
this, verify that there is at least one vacant slot in the vector.

The implementation allows for repeated probing without setting up the
stack frame should calls to eq not be necessary.