hashtab: Handle edge case where vacant and probe end not in same vector
This addresses the case where a vector contains a vacant slot but
not a probe end. When this occurs, the offset of the vacant slot
is now saved while the probing continues until a probe end is
encountered. If not matches are found before the probe end, the
index that was set aside is returned.