hashtab: Handle edge case triggering early probe stop on match
The issue occurs when the vector contains both at least one vacant
and one slot with matching hash. In this case, the vacant slot
would be returned if none of the potentially matching slots were
actually matching, regardless of whether the vector contained a
probe end of not.

This ensures the probing is carried out until the end regardless