btree: Ignore supposed null dereference
The interesting path is

/home/vilhelm/repos/c/scc/src/scc_btree.c:859:5: note: 'sibling'
initialized to a null pointer value
    struct scc_btnode_base *sibling = 0;
    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/vilhelm/repos/c/scc/src/scc_btree.c:860:8: note: Assuming 'bound'
is 0
    if(bound) {
       ^~~~~
/home/vilhelm/repos/c/scc/src/scc_btree.c:860:5: note: Taking false
branch
    if(bound) {
    ^
/home/vilhelm/repos/c/scc/src/scc_btree.c:868:8: note: 'bound' is >=
field 'bt_nkeys'
    if(bound < curr->bt_nkeys) {
       ^~~~~
/home/vilhelm/repos/c/scc/src/scc_btree.c:868:5: note: Taking false
branch
    if(bound < curr->bt_nkeys) {
    ^
/home/vilhelm/repos/c/scc/src/scc_btree.c:879:39: note: Passing null
pointer value via 3rd parameter 'sibling'
    scc_btnode_merge_left(base, next, sibling, curr, bound, elemsize);

i.e. it occurs only if bound == 0 and bound >= curr->bt_nkeys which,
as bt_nkeys is unsigned, implies that curr->bt_nkeys == 0. This is
possible only in the root of an empty tree. However, insertion in
an empty tree never, under any circumstance, attempts to balance the
tree. Hence, this is a false positive.