btmap: Avoid UB when initializing btmap
N2350 kicked up quite the discussion. Seems it culminated in the
consensus that defining types in offsetof pre-C23 is UB. Since scc
aims for C99-compatibility, it can't rely on 23 extensions. Implement
it using flexible array members instead.

This uses slightly more memory but should, in the grand scheme of
things, be negligible